Update Stagehand Sample App to be compatible with Stagehand v3 (#62)
* Update Stagehand Sample App to Stagehand v3

Updated Stagehand example app to v3 Stagehand compatible version of script.

* Remove -v3 in readme

Remove "-v3" from the end of the app name as it's not used.

* Update next steps for stagehand template

Updated the next steps invocation example for Stagehand so that it's accurate now.

* Update input/output schemas and payloads

Add more descriptive and accurate input and output schemas and payloads for the stagehand sample app.